deleted duplicated fragments for datasets
